floorf instead of floor

+++ src/cpp/cuda/fast_splat_2d_cuda.cu
@@ -141,7 +141,7 @@ bilinear_splat(const float src_red, const float src_green, const float src_blue,
                const float x_in_tile, const float y_in_tile, float *tile) {
   const int left = floorf(x_in_tile);
   const int right = left + 1;
-  const int top = floor(y_in_tile);
+  const int top = floorf(y_in_tile);
   const int bottom = top + 1;
 
   if (left >= 0 && left < TILE_SIZE_X) {
